+++ Dockerfile
@@ -1,4 +1,3 @@
-FROM eclipse-temurin:17-jdk-alpine
-VOLUME /tmp
+FROM eclipse-temurin:21-jdk-alpine
 COPY target/*.jar app.jar
 ENTRYPOINT ["java","-jar","/app.jar"]
