+++ src/main/java/io/orkes/demo/banking/controller/BankingApiController.java
@@ -3,7 +3,7 @@ package io.orkes.demo.banking.controller;
 import io.orkes.demo.banking.pojos.DepositDetail;
 import io.orkes.demo.banking.service.FraudCheckService;
 import io.orkes.demo.banking.service.WorkflowService;
-import io.orkes.demo.banking.workers.FraudCheckResult;
+import io.orkes.demo.banking.pojos.FraudCheckResult;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
@@ -17,9 +17,7 @@ import java.util.Map;
 @AllArgsConstructor
 @RestController
 public class BankingApiController {
-
     private final FraudCheckService fraudCheckService;
-
     private final WorkflowService workflowService;
 
     @PostMapping(value = "/checkForFraud", produces = "application/json")
@@ -28,14 +26,11 @@ public class BankingApiController {
         return ResponseEntity.ok(fraudCheckService.checkForFraud(depositDetail));
     }
 
-
     // docs-marker-start-1
     @PostMapping(value = "/triggerDepositFlow", produces = "application/json")
     public ResponseEntity<Map<String, Object>> triggerDepositFlow(@RequestBody DepositDetail depositDetail) {
         log.info("Starting deposit flow for: {}", depositDetail);
         return ResponseEntity.ok(workflowService.startDepositWorkflow(depositDetail));
     }
-
     // docs-marker-end-1
-
 }
