+++ src/main/java/io/orkes/demo/banking/pojos/DepositDetail.java
@@ -1,19 +1,14 @@
 package io.orkes.demo.banking.pojos;
 
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
 import java.math.BigDecimal;
 
 @Data
-@Builder
-@NoArgsConstructor
 @AllArgsConstructor
+@NoArgsConstructor
 public class DepositDetail {
-
     private String accountId;
     private BigDecimal amount;
-
 }
