+++ src/main/java/io/orkes/demo/banking/pojos/FraudCheckResult.java
@@ -1,14 +1,12 @@
-package io.orkes.demo.banking.workers;
+package io.orkes.demo.banking.pojos;
 
 import lombok.Data;
 
 @Data
 public class FraudCheckResult {
-
     public enum Result {
         PASS, FAIL;
     }
-
     private Result result;
     private String reason;
 }
