+++ src/main/java/io/orkes/demo/banking/service/FraudCheckService.java
@@ -1,20 +1,17 @@
 package io.orkes.demo.banking.service;
 
+import static io.orkes.demo.banking.pojos.FraudCheckResult.Result.FAIL;
+import static io.orkes.demo.banking.pojos.FraudCheckResult.Result.PASS;
 import io.orkes.demo.banking.pojos.DepositDetail;
-import io.orkes.demo.banking.workers.FraudCheckResult;
+import io.orkes.demo.banking.pojos.FraudCheckResult;
 import org.springframework.stereotype.Service;
-
 import java.math.BigDecimal;
-
-import static io.orkes.demo.banking.workers.FraudCheckResult.Result.FAIL;
-import static io.orkes.demo.banking.workers.FraudCheckResult.Result.PASS;
 
 @Service
 public class FraudCheckService {
-
     public FraudCheckResult checkForFraud(DepositDetail depositDetail) {
         FraudCheckResult fcr = new FraudCheckResult();
-        if(depositDetail.getAmount().compareTo(BigDecimal.valueOf(100000)) > 0) {
+        if (depositDetail.getAmount().compareTo(BigDecimal.valueOf(100_000)) > 0) {
             fcr.setResult(FAIL);
             fcr.setReason("Amount too large");
         } else {
@@ -23,5 +20,4 @@ public class FraudCheckService {
         }
         return fcr;
     }
-
 }
