+++ src/main/resources/application.properties
@@ -1,9 +1,3 @@
-
-# swagger-ui custom path
-springdoc.swagger-ui.path=/swagger-ui.html
-management.endpoints.enabled-by-default=false
-management.endpoint.info.enabled=false
-server.port=8081
 # Obtain key and secret by logging into https://play.orkes.io/
 # and navigating to applications menu, create an application and generate key/secret
 conductor.security.client.key-id=
