--- src/test/java/io/orkes/demo/banking/BankingApplicationTests.java
@@ -1,14 +1,0 @@
-package io.orkes.demo.banking;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BankingApplicationTests {
-
-	@Test
-	void contextLoads() {
-
-	}
-
-}
